+++ src/multi_vector_simulator/utils/constants_json_strings.py
@@ -229,18 +229,25 @@ LCOE_ASSET = "levelized_cost_of_energy_of_asset"
 # Other KPI
 TOTAL_RENEWABLE_GENERATION_IN_LES = "Total internal renewable generation"
 TOTAL_NON_RENEWABLE_GENERATION_IN_LES = "Total internal non-renewable generation"
+TOTAL_GENERATION_IN_LES = "Total internal generation"
 TOTAL_RENEWABLE_ENERGY_USE = "Total renewable energy use"
 TOTAL_NON_RENEWABLE_ENERGY_USE = "Total non-renewable energy use"
 RENEWABLE_FACTOR = "Renewable factor"
 RENEWABLE_SHARE_OF_LOCAL_GENERATION = "Renewable share of local generation"
 
+
 TOTAL_DEMAND = "Total_demand"
 TOTAL_EXCESS = "Total_excess"
+TOTAL_FEEDIN = "Total_feedin"
 SUFFIX_ELECTRICITY_EQUIVALENT = "_electricity_equivalent"
 ATTRIBUTED_COSTS = "Attributed costs"
 LCOeleq = "Levelized costs of electricity equivalent"
 
 DEGREE_OF_SECTOR_COUPLING = "Degree of sector coupling"
+DEGREE_OF_AUTONOMY = "Degree of autonomy"
+ONSITE_ENERGY_FRACTION = "Onsite energy fraction"
+ONSITE_ENERGY_MATCHING = "Onsite energy matching"
+
 
 # KPI_FLOW_MATRIX
 KPI_SCALARS = (